// src/types.ts
export interface SocialLink {
  platform: SocialPlatform
  handle: string
}

export type SocialPlatform =
  | 'INSTAGRAM'
  | 'X'
  | 'BLUESKY'
  | 'LINKEDIN'
  | 'YOUTUBE'
  | 'SPOTIFY'
  | 'TIKTOK'
  | 'SNAPCHAT'
  | 'GITHUB'
  | 'FACEBOOK'
  | 'REDDIT'
  | 'DISCORD'
  | 'TWITCH'
  | 'TELEGRAM'
  | 'PINTEREST'
  | 'TUMBLR'
  | 'SOUNDCLOUD'
  | 'BANDCAMP'
  | 'PATREON'
  | 'KO_FI'
  | 'MASTODON'
  | 'WEBSITE'
  | 'EMAIL'